pyproject.toml
[project]
name = "project-status-log-changed"
version = "0.1.0"
description = "Automated tracking and reporting system for project status changes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "faker>=25.0.0",
    "python-dateutil>=2.9.0"
]
